pyproject.toml
[project]
name = "routina"
version = "0.1.0"
description = "App de coaching de fitness asistida por IA — caso de estudio del programa de evals (alaimolabs.com/es/eva)"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.30,<1",
    "openai>=1.55,<2",
    "jsonschema>=4.23,<5",
    "python-dotenv>=1.0,<2",
]

[tool.uv]
package = false
